// src/openapi/models/document-space-app-client-response-dto-wrapper.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TRON Common API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { DocumentSpaceAppClientResponseDto } from './document-space-app-client-response-dto';

/**
 * 
 * @export
 * @interface DocumentSpaceAppClientResponseDtoWrapper
 */
export interface DocumentSpaceAppClientResponseDtoWrapper {
    /**
     * 
     * @type {Array<DocumentSpaceAppClientResponseDto>}
     * @memberof DocumentSpaceAppClientResponseDtoWrapper
     */
    data: Array<DocumentSpaceAppClientResponseDto>;
}